Dto e AutoMapper

+++ Cwk.Booking.Api/Controllers/HotelsController.cs
@@ -1,21 +1,17 @@
-﻿using CwkBooking.Dal;
+﻿using AutoMapper;
+using CwkBooking.Api.Dtos;
+using CwkBooking.Dal;
 using CwkBooking.Domain.Models;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
-using System;
+using Microsoft.Extensions.Logging;
 using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 
 namespace CwkBooking.Api.Controllers
 {
 
-    //CRUD
-    //Create
-    //Read - get all, get by id
-    //Update
-    //Delete
-    // /hotels
     [ApiController]
     [Route("api/[controller]")]
     public class HotelsController : Controller
@@ -23,19 +19,23 @@ namespace CwkBooking.Api.Controllers
         private readonly ILogger<HotelsController> _logger;
         private readonly HttpContext _http;
         private readonly DataContext _ctx;
+        private readonly IMapper _mapper;
         public HotelsController(ILogger<HotelsController> logger, IHttpContextAccessor httpContextAccessor,
-            DataContext ctx)
+            DataContext ctx, IMapper mapper)
         {
             _logger = logger;
             _http = httpContextAccessor.HttpContext;
             _ctx = ctx;
+            _mapper = mapper;
         }
 
         [HttpGet]
         public async Task<IActionResult> GetAllHotels()
         {
             var hotels = await _ctx.Hotels.ToListAsync();
-            return Ok(hotels);
+            var hotelsGet = _mapper.Map<List<HotelGetDto>>(hotels);
+
+            return Ok(hotelsGet);
         }
 
 
@@ -44,28 +44,37 @@ namespace CwkBooking.Api.Controllers
         public async Task<IActionResult> GetHotelById(int id)
         {
             var hotel = await _ctx.Hotels.FirstOrDefaultAsync(h => h.HotelId == id);
-            return Ok(hotel);
+
+            if (hotel == null)
+                return NotFound();
+
+            var hotelGet = _mapper.Map<HotelGetDto>(hotel);
+            return Ok(hotelGet);
         }
 
         [HttpPost]
-        public async Task <IActionResult> CreateHotel([FromBody] Hotel hotel)
+        public async Task<IActionResult> CreateHotel([FromBody] HotelCreateDto hotel)
         {
-            _ctx.Hotels.Add(hotel);   //save in memory
-            await _ctx.SaveChangesAsync();  // save DataBase
-            return CreatedAtAction(nameof(GetHotelById), new {id = hotel.HotelId}, hotel);
+            var domainHotel = _mapper.Map<Hotel>(hotel);
+
+            _ctx.Hotels.Add(domainHotel);
+            await _ctx.SaveChangesAsync();
+
+            var hotelGet = _mapper.Map<HotelGetDto>(domainHotel);
+
+            return CreatedAtAction(nameof(GetHotelById), new { id = domainHotel.HotelId }, hotelGet);
         }
 
         [HttpPut]
         [Route("{id}")]
-        public async Task<IActionResult> UpdateHotel([FromBody] Hotel updated, int id)
+        public async Task<IActionResult> UpdateHotel([FromBody] HotelCreateDto updated, int id)
         {
-            var hotel = await _ctx.Hotels.FirstOrDefaultAsync(h =>h.HotelId == id);
-            hotel.Stars = updated.Stars;
-            hotel.Name = updated.Name;
-            hotel.Description = updated.Description;
-            
-            _ctx.Hotels.Update(hotel);
+            var toUpdate = _mapper.Map<Hotel>(updated);
+            toUpdate.HotelId = id;
+
+            _ctx.Hotels.Update(toUpdate);
             await _ctx.SaveChangesAsync();
+
             return NoContent();
         }
 
@@ -74,8 +83,13 @@ namespace CwkBooking.Api.Controllers
         public async Task<IActionResult> DeleteHotel(int id)
         {
             var hotel = await _ctx.Hotels.FirstOrDefaultAsync(h => h.HotelId == id);
+
+            if (hotel == null)
+                return NotFound();
+
             _ctx.Hotels.Remove(hotel);
             await _ctx.SaveChangesAsync();
+
             return NoContent();
         }
     }
